+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 [Full Changelog](http://github.com/myronmarston/vcr/compare/v1.10.0...master)
 
 * Fix typo in error message [Bradley](https://github.com/bradleyisotope).
+* Fix excon adapter to properly handle queries specified as a hash.
 
 ## 1.10.0 (May 18, 2011)